pyproject.toml
[project]
name = "camera-manual-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.1",
    "uvicorn>=0.34.2",
]
